Restructure CLI and configuration
This commit reworks the way the CLI is constructed, moving all relevant
files into the cmd/vpc package structure. It also more cleanly separates
the configuration, and uses `viper.Unmarshal` instead of `viper.GetX`.

Database configuration is moved to the `db` package, and subcommands of
`db` now use a database pool directly instead of starting an agent.

Fatal log messages are now hooked and the `conswriter` is closed, which
should reset a terminal to the previous state.